fix cycle, refactor useSelectionZone
